+++ source/vm_video_transcoder/lib/plugin_stream_mapper.py
@@ -85,22 +85,21 @@ class PluginStreamMapper(StreamMapper):
             if advanced_options:
                 # Overwrite all advanced options
                 self.advanced_options = advanced_options
+            # Add metadata tag AFTER overwrite so it's never lost
+            self.set_ffmpeg_advanced_options(**{'-metadata': 'unmanic_status=processed'})
+            tools.append_worker_log(self.worker_log, "Adding metadata tag 'unmanic_status=processed'")
             # Don't apply any other settings
             return
 
         # Build default options of standard mode
         if self.settings.get_setting('mode') == 'standard':
 
-            # Set max muxing queue size and metadata
+            # Set max muxing queue size
             advanced_kwargs = {}
             if self.settings.get_setting('max_muxing_queue_size'):
                 advanced_kwargs['-max_muxing_queue_size'] = str(self.settings.get_setting('max_muxing_queue_size'))
-            
-            # Add metadata tag automatically
-            advanced_kwargs['-metadata'] = 'unmanic_status=processed'
-            
+
             self.set_ffmpeg_advanced_options(**advanced_kwargs)
-            tools.append_worker_log(self.worker_log, "Adding metadata tag 'unmanic_status=processed'")
 
         # Check for config specific settings in modes that expose smart filters
         if self.settings.get_setting('mode') in ['basic', 'standard']:
@@ -119,6 +118,10 @@ class PluginStreamMapper(StreamMapper):
             generic_kwargs, advanced_kwargs = encoder_lib.generate_default_args()
             self.set_ffmpeg_generic_options(**generic_kwargs)
             self.set_ffmpeg_advanced_options(**advanced_kwargs)
+
+        # Add metadata tag LAST for basic/standard modes so no encoder can overwrite it
+        self.set_ffmpeg_advanced_options(**{'-metadata': 'unmanic_status=processed'})
+        tools.append_worker_log(self.worker_log, "Adding metadata tag 'unmanic_status=processed'")
 
     def enable_execution_stage(self):
         """